Fix CADD_S_H CADD_D_H tests

+++ core/unit_test/TestHalfOperators.hpp
@@ -430,7 +430,7 @@ struct Functor_TestHalfOperators {
     expected_lhs(CADD_H_S) = d_lhs;
     expected_lhs(CADD_H_S) += d_rhs;
 
-    tmp_s_lhs = h_lhs;
+    tmp_s_lhs = static_cast<float>(h_lhs);
     tmp_s_lhs += h_rhs;
     actual_lhs(CADD_S_H) = static_cast<double>(tmp_s_lhs);
     expected_lhs(CADD_S_H) = d_lhs;
@@ -442,7 +442,7 @@ struct Functor_TestHalfOperators {
     expected_lhs(CADD_H_D) = d_lhs;
     expected_lhs(CADD_H_D) += d_rhs;
 
-    tmp_d_lhs = h_lhs;
+    tmp_d_lhs = static_cast<double>(h_lhs);
     tmp_d_lhs += h_rhs;
     actual_lhs(CADD_D_H) = static_cast<double>(tmp_d_lhs);
     expected_lhs(CADD_D_H) = d_lhs;
